refactor: 상품 API 변경
상품이 상점을 변경할 수 없도록 수정
예외 메시지 통일
삭제 시 존재하지 않는 엔티티에 대하여 예외 처리 추가

+++ productService/src/main/java/ecommerce/productService/product/service/StoreService.java
@@ -30,16 +30,19 @@ public class StoreService {
 	}
 
 	public Store findById(Long id) {
-		return storeRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Store not found."));
+		return storeRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Store with ID " + id + " not found."));
 	}
 
 	public Store updateStore(Long id, Store updateStore) {
-		Store store = storeRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Store not found."));
+		Store store = storeRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Store with ID " + id + " not found."));
 		store.update(updateStore);
 		return store;
 	}
 
 	public void deleteStore(Long id) {
+		if(!storeRepository.existsById(id)) {
+			throw new EntityNotFoundException("Store with ID " + id + " not found.");
+		}
 		storeRepository.deleteById(id);
 	}
 }
